Számlára csak a bankból: minden jutalom fizikai veretben fizet
- CurrencyManager.payOutTokens: veret-itemek a kézbe (64-es kötegek, túlcsordulás a földre)
- Kopott erszény beváltása fizikai vereteket ad (nem egyenleg-jóváírást)
- Felvásárló veretben fizet, egészre lefelé kerekítve (1 veret alatt nem vesz)
- Quest, napi/heti kihívás, mérföldkő, parkour, ambient-esemény, AFK-jutalom, vérdíj (killer-szál hop) és Bankbetét-jegy mind token-itemben fizet
- Piac/aukció/kincstár bankon belüli átvezetés marad; a számlára kizárólag /bank deposit tesz pénzt
- PLAYTEST + player-guide 03 frissítve

+++ src/main/java/hu/taliann/icesmp/commands/ItemGiveCommand.java
@@ -124,15 +124,15 @@ public final class ItemGiveCommand implements BasicCommand {
             case "erszeny" -> {
                 // Kopott erszény: az <id> itt az ÖSSZEG, a [darab] az erszények száma,
                 // a valuta erszényenként véletlen (a mob-drop/horgász-lelet útjával azonos).
-                final double value;
+                final long value;
                 try {
-                    value = Double.parseDouble(id);
+                    value = Long.parseLong(id);
                 } catch (final NumberFormatException exception) {
                     sender.sendMessage(messageManager.get("admin.iceitem.bad-amount",
                             "&cÉrvénytelen összeg: &f%s", id));
                     return;
                 }
-                if (value <= 0.0D || moneyPouchFactory == null) {
+                if (value <= 0L || moneyPouchFactory == null) {
                     sender.sendMessage(messageManager.get("admin.iceitem.bad-amount",
                             "&cÉrvénytelen összeg: &f%s", id));
                     return;
